Changed int args to float

+++ code_0.cpp
@@ -3,7 +3,7 @@ int add(int a, int b)
   return a+b;
 }
 
-float multiply(int a, int b)
+float multiply(float a, float b)
 {
   return a*b;
 }
